Add MultiDropdown and Refacto
