+++ shared/@hedvig-ui/Modal/index.tsx
@@ -1,28 +1,9 @@
 import styled from '@emotion/styled'
 import { FadeIn } from '@hedvig-ui'
-import React, { useEffect, useRef } from 'react'
+import React, { useRef } from 'react'
 import { X as CloseIcon } from 'react-bootstrap-icons'
 import { Portal } from 'react-portal'
-
-export const useClickOutside = (ref: any, handler: (e: any) => void) => {
-  useEffect(() => {
-    const listener = (event: any) => {
-      if (!ref.current || ref.current.contains(event.target)) {
-        return
-      }
-
-      handler(event)
-    }
-
-    document.addEventListener('mousedown', listener)
-    document.addEventListener('touchstart', listener)
-
-    return () => {
-      document.removeEventListener('mousedown', listener)
-      document.removeEventListener('touchstart', listener)
-    }
-  }, [ref, handler])
-}
+import { useClickOutside } from '../utils/click-outside'
 
 const ModalWrapperStyled = styled.div<{
   position?: 'top' | 'center' | 'bottom'
