+++ shared/@hedvig-ui/utils/click-outside.ts
@@ -1,0 +1,21 @@
+import { useEffect } from 'react'
+
+export const useClickOutside = (ref: any, handler: (e: any) => void) => {
+  useEffect(() => {
+    const listener = (event: any) => {
+      if (!ref.current || ref.current.contains(event.target)) {
+        return
+      }
+
+      handler(event)
+    }
+
+    document.addEventListener('mousedown', listener)
+    document.addEventListener('touchstart', listener)
+
+    return () => {
+      document.removeEventListener('mousedown', listener)
+      document.removeEventListener('touchstart', listener)
+    }
+  }, [ref, handler])
+}
